Record application id at configuration phase
This change removes AnalyticsRecordingTask which records application id
at execution time. It is nice to remove that task because that task is
set to always run which looks confusing to users.

As the application id obtained in dynamice feature module is from a task
output of the application module, so we couldn't record application id
from dynamic feature at configuration time.

Fixes: 269758603
Test: ProfileContentTest
Change-Id: I89680a8688203d5aaa5166eb860be7b056349ea2