New Module class
Used to represent a (group, name) pair; sometimes also referred to
as (groupId, artifactId).  Modules can be parsed from strings
containing exactly one colon character; an identifier can be generated
from a Module if neither the group nor the name contains a colon.

The name `Module` is perhaps unfortunate, given the number of other
Module concepts floating around the codebase which have very little to
do with this Module concept; however, it is taken directly from the
Gradle lexicon.

Bug: 242691473
Test: added
Change-Id: Ie080be09430d841ce179ba69705def6cc1667816