Use Directory.getOrderedFileTree() to ensure stable order
Currently, Gradle's FileTree doesn't guarantee stable order, which leads
to non-deterministic task outputs and cache misses (see
https://github.com/gradle/gradle/issues/21379 for more details).

To work around this issue, we provide a utility extension function
called `Directory.getOrderedFileTree()` which returns all regular files
in the directory in stable order (even across platforms).

In this commit, we use `Directory.getOrderedFileTree()` on the task
inputs of MergeGeneratedProguardFiles.

We can use it in more places later if the Gradle bug above is not fixed
soon.

Bug: 266403349
Test: Added GradleApiUtilsTest
Change-Id: I6de99e88331befe9812bd3d2d6160076de2333f1