A prefixInfimum is not (necessarily) a preview version
No actual piece of software (Component, in Gradle terminology) will
have a prefixInfimum version.  So this should not change any behaviour
relating to filtering out components.  Prefix infima are generally
used to mark boundaries of ranges, for example in the implementation
of prefix and some maven-style version ranges.

A question that will be asked of those ranges is "do they involve
previews".  In that sense it is useful to be able to distinguish
between the range denoted by 1.0.+ and the range denoted by 1.alpha.+
-- in the latter, there is an explicit preview signifier (the
non-numeric Part) and in the former, there isn't.  However, both
ranges are represented internally with a lower bound of a prefix
infimum, so an attempt to distinguish between these classes of ranges
would fail if prefix infima are treated as preview versions.

A theoretical justification for this change is that the prefix infimum
version 2.dev.dev.dev... is actually equivalent to the prefix supremum
1.∞.∞.∞... (in the sense that 1.999... is equivalent to 2.000...).
The fact that we use prefix infima, with their infinite extent of
non-numeric parts, is an implementation artifact, not a fundamental
attribute of that point in the extended Version space.

Bug: 242691473
Test: existing, added
Change-Id: I9a0d575fecb5c14a64a6c8b17c15f1186412ff6e