Use appropriate entity representation in GoogleMavenRepository
The previous implementation works in terms of GradleCoordinate, which
unfortunately does not allow for an explicit separation between
wildcards or ranges and fully resolved entities.  In this new
representation, we query a repository with a dependency specifier, and
get back one or a sequence of appropriate Components.

We preserve existing semantics for now, even though they're not
desperately well-founded: if the Dependency that we query with has an
explicit non-singleton, we treat that as a strict requirement (and
hence use .contains()); if it is a singleton, then we use .accepts().

This change propagates into aspects of the Project System (in the
linked change in tools/adt/idea); the consequence is that we can
provide greater clarity not only here but in the analysis of
dependency specifier compatibility.

Extension functions defined in this change are extracted, documented
and tested in a subsequent CL.

Bug: 242691473
Bug: 279886738
Test: existing tests adjusted
Change-Id: I3eb22afac32948aed3d10dbc06db5253fb77b076