Add a convenience routine to VersionRange
Gradle's Rich Version semantics involve weakening the upper bound of
non-strict dependency versions.  We will need this in our
implementation of RichVersion, in the next CL, so define, implement
and test withoutUpperBound().  For now keep it as an internal
function, because this should only be used in the implementation of
RichVersion, though in principle we could make this public.

Bug: 242691473
Test: added
Change-Id: I7f27ba41e9c9437f3ad62c62fddbb584500d5208