Define and test singleton queries for VersionRange
Some operations on RichVersion and Dependency will want to do
different things depending on whether the rich version expresses a
singleton (required or strict) dependency, because that in practice
tends to express a stronger requirement than a range even if in theory
there is no strong distinction between a singleton and a non-singleton
range in terms of dependency resolution.

Define isSingleton and singletonVersion with pragmatic
semantics (i.e. non-singleton) for prefixInfima.  Document and test
them.

Bug: 242691473
Test: added
Change-Id: Ifc252e62393766e30440ccfab1c3fc232ec220fc