Improve test stability by introducing a forced delay
Previously, the test would sometimes execute so fast that the creation
timestamp was the same as the updated timestamp. This caused the test to
fail, as the timestamp change was not detected.

To fix this, we have introduced a forced delay in the test. This delay
is small enough to not affect the performance of the test, but large
enough to ensure that the timestamp change is detected.

Bug: 277719430
Test: Unit test was updated.
Change-Id: I70d3901a7c259aebf529e4db86679c37a77949e5