Add artifact property to LintModelVariant (and deprecate mainArtifact)
This change adds a new artifact property to LintModelVariant and
deprecates the mainArtifact property.

Instead of renaming mainArtifact to artifact, mainArtifact is deprecated
to avoid breaking existing 3rd party lint checks that use it.

This change also adds a type property to LintModelArtifact.

This change is motivated by the plan to migrate to doing lint
analysis per artifact when running lint from AGP.

Bug: 197322928
Test: existing
Change-Id: I20a10f3fdf94a696f8334e870d11bee476a4b340