Make Version implement Serializable
It is a little more complex than just claiming to implement the
interface.  We have two possible representations for our
serialization: we could serialize the string representation (and the
prefixInfimum boolean), treating the parts/separators lists as
transient; this would have the advantage that we do not have to worry
about serializing Part and Separator objects, but involves writing
writeObject/readObject methods.  The alternative is to
serialize the object directly; from the point of view of the Version
class itself, this is simpler, but involves changes to internals.

So, I tried the first one, and it essentially worked as far as direct
serialization went.  However, the IDE's use of Bean serialization
imposed some additional restrictions; in particular, it seems like
there are extra constraints involved in serializing through that
mechanism, and fewer customization hooks.  So, turning to the
alternative, we observe that Separator is already serializable and
turn to Part, which is broadly serializable, except: BigIntegers in
the Int range are serialized as Int, and consequently lead to
exceptions when attempting to pass them to a constructor that accepts
BigInteger.

So, the constructor cannot have a BigInteger argument.  And, because
of the way the IDE uses the serialization infrastructure, we can't
have a BigInteger field, or at least not a non-nullable one, because
the instance constructed will have a null value for that field.  We
therefore have to implement lazy initialization of that property, and
we can't use `by lazy` because the synthetic class for the delegating
lambda is also not serializable.  So implement it by hand.

While we're at it, take the opportunity to simplify slightly the
Special classes, where we can use a property rather than relying on
the caller to initialize them correctly.

Bug: 279886738
Test: added
Change-Id: I3f392c40ea7734095ac65ac64f2903c7e5d80ac6