Add RISCV support to AGP
This adds basic support for riscv ABI in Android Gradle Plugin.
This is based on an ad hoc NDK r26 that has riscv cobbled in.
Not all build systems worked (at least CMake 3.10) didn't work,
so it wasn't possible to get all tests passing.

As part of this, I made it possible to add an ABI just by
publishing it in meta/abis.json without also adding it to
the Abi enum. This only works for ABIs that only exist in
NDKs that support meta/abis.json.

Bug: NA
Test: Added tests for riscv64. Adjusted existing.

Change-Id: I61afa0a144b9b5f0250ef2431f709020905916cd