Provide the preview supremum as well as the infimum
This allows us to more rigorously compute the notion of the preview
"series": all those versions between the infimum and the supremum.
Use this in the GradleDetector dependency version filter.

In the process, also fix some internal representation issues: the
previewPrefix (now renamed previewInfimum) computation could cause
unexpected results in two ways: the last separator was not EMPTY, and
the previewPrefix of a version whose first element was non-Numeric had
no fields.  Fix those representation issues, and institute checks
against them in the (private) constructor.

Bug: 242691473
Test: added Version unit tests; existing GradleDetectorTest
Change-Id: I105d5409af41ab4f37b08ca898f415fe088a70b5