Add application context checks where they were missing
To fix b/206207283 in LeakDetector, ag/16276990 added checks for whether
a Context field is an application context. These checks look for the
name 'appContext' and for the @ApplicationContext annotation, as well
as assignment to getApplicationContext().

But these checks are only performed when static fields are directly
visited, and not when fields of a static class are visited starting
from visitClass. This CL adds the same checks in that case.

Bug: 262841842
Test: testHilt, testLifeCycle
Change-Id: Ia546552441e28c95bc3849008ecb1ca2c51bcdaa